Bug 1181572 - Store job durations in the new unified table using the ORM
Instead of storing the calculated average job durations in the
per-project job_eta tables, they are now stored in the new job_duration
table in the main treeherder database. This means we can use Django's
ORM instead of datasource.

In addition, if a job signature already exists in the table, we update
it rather than creating duplicate rows - and most importantly unlike the
previous table, the correct columns are indexed to avoid performance
problems.

This commit just changes where we store new durations, a later commit
will switch the retrieval of them from the old table to the new.